Resolve asm operand width warnings
Clang is a bit more pedantic about operand widths, so read first into a 64 bit value and cast to 32 bits.

+++ include/sddf/benchmark/sel4bench.h
@@ -76,9 +76,9 @@ static FASTFN void sel4bench_private_write_pmcr(uint32_t val)
 
 static FASTFN uint32_t sel4bench_private_read_pmcr(void)
 {
-    uint32_t val;
+    uint64_t val;
     PMU_READ(PMCR, val);
-    return val;
+    return (uint32_t)val;
 }
 
 #define MODIFY_PMCR(op, val) sel4bench_private_write_pmcr(sel4bench_private_read_pmcr() op (val))
@@ -90,9 +90,9 @@ static FASTFN void sel4bench_private_write_cntens(uint32_t mask)
 
 static FASTFN uint32_t sel4bench_private_read_cntens(void)
 {
-    uint32_t mask;
+    uint64_t mask;
     PMU_READ(PMCNTENSET, mask);
-    return mask;
+    return (uint32_t)mask;
 }
 
 static FASTFN void sel4bench_private_write_cntenc(uint32_t mask)
@@ -102,9 +102,9 @@ static FASTFN void sel4bench_private_write_cntenc(uint32_t mask)
 
 static FASTFN uint32_t sel4bench_private_read_pmcnt(void)
 {
-    uint32_t val;
+    uint64_t val;
     PMU_READ(PMXEVCNTR, val);
-    return val;
+    return (uint32_t)val;
 }
 
 static FASTFN void sel4bench_private_write_pmcnt(uint32_t val)
